requirements.txt
backcall==0.1.0
cached-property==1.5.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==2.1.7
django-dotenv==1.4.2
django-filter==2.1.0
django-url-filter==0.3.12
djangorestframework==3.9.1
enum-compat==0.0.2
idna==2.8
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7